Add files via upload
Hi everyone I am excited to share my latest Project - Road Accident Analysis Dashboard in Power BI.

⭐ 📊 Road Accident Analysis dashboard from Power BI! 📊 ⭐

 🎉This is the Advance Dashboard with lot of cool methods🎉

Road accidents are a significant concern for public safety and policy makers. To better understand the factors contributing to accidents and identify patterns, we conducted a data analysis using Power BI. In this project, we delved into a comprehensive dataset on road accidents, examining trends, and gaining insights to drive informed decisions for safer roads.

This dashboard is having a complete dynamic & interactive dashboard which will be operated by few Filters and few Slicers. I've also used some iteractive and quick filters as well.This dashboard is completely dynamic on the basis of clients requirement in which they want to create a road accident  dashboard for year 2021 to 2022 so they can have insight on the below requirement -

1- Primary KPI -
-- Total Casualties and total accident values for Current Year and YoY growth.
-- Total Casualties by Accident Severity for current year and YoY growth.

2-Secondary KPI -
--Total Casualties with respect to vehicle type for Current Year.

3- Monthly Trend showing comparison of casualties for current year and previous year.
4- Casualties by Road Type for current year
5- Current year casualties by Area/Location & Day/Night
6- Total casualties and total Accident by Location

The Power BI Functionalities I used are as following-

--Data cleaning in Power Query
--Data processing
--Time intelligence function/ Calendar Date table in Power BI
--Data modelling (Relationship b/w multiple tables)
--YTD and YoY Growth Calculations using DAX
--KPI and Advanced KPI Generations
--Creating custom columns and measures in the reports
--Importing Images
--Creating different charts and generating insights

#powerbi #dataanalytics #roadaccident #analysis #dashboard